+++ data/db-room/src/commonMain/kotlin/io/github/alelk/pws/database/book/BookDetailProjection.kt
@@ -1,0 +1,28 @@
+package io.github.alelk.pws.database.book
+
+import androidx.room.ColumnInfo
+import io.github.alelk.pws.domain.core.Locale
+import io.github.alelk.pws.domain.core.Version
+import io.github.alelk.pws.domain.core.Year
+import io.github.alelk.pws.domain.core.ids.BookId
+import io.github.alelk.pws.domain.core.ids.SongNumberId
+import io.github.alelk.pws.domain.person.Person
+
+data class BookDetailProjection(
+  @ColumnInfo(name = "id") val id: BookId,
+  @ColumnInfo(name = "version") val version: Version,
+  @ColumnInfo(name = "locale") val locale: Locale,
+  @ColumnInfo(name = "name") val name: String,
+  @ColumnInfo(name = "display_short_name") val displayShortName: String,
+  @ColumnInfo(name = "display_name") val displayName: String,
+  @ColumnInfo(name = "release_date") val releaseDate: Year? = null,
+  @ColumnInfo(name = "authors") val authors: List<Person>? = null,
+  @ColumnInfo(name = "creators") val creators: List<Person>? = null,
+  @ColumnInfo(name = "reviewers") val reviewers: List<Person>? = null,
+  @ColumnInfo(name = "editors") val editors: List<Person>? = null,
+  @ColumnInfo(name = "description") val description: String? = null,
+  @ColumnInfo(name = "preface") val preface: String? = null,
+  @ColumnInfo(name = "count_songs") val countSongs: Int,
+  @ColumnInfo(name = "first_song_number_id") val firstSongNumberId: SongNumberId?,
+  @ColumnInfo(name = "priority") val priority: Int,
+)