+++ data/db-room/src/commonMain/kotlin/io/github/alelk/pws/database/bookstatistic/BookStatisticDao.kt
@@ -1,4 +1,4 @@
-package io.github.alelk.pws.database.book_statistic
+package io.github.alelk.pws.database.bookstatistic
 
 import androidx.room.Dao
 import androidx.room.Delete
@@ -7,7 +7,7 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Transaction
 import androidx.room.Upsert
-import io.github.alelk.pws.database.book_statistic.BookStatisticWithBookEntity
+import io.github.alelk.pws.domain.bookstatistic.query.BookStatisticQuery
 import io.github.alelk.pws.domain.core.ids.BookId
 import kotlinx.coroutines.flow.Flow
 
@@ -20,10 +20,10 @@ interface BookStatisticDao {
   suspend fun insert(bookStatistics: List<BookStatisticEntity>): List<Long>
 
   @Upsert
-  suspend fun update(bookStatistics: List<BookStatisticEntity>): List<Long>
+  suspend fun upsert(bookStatistics: List<BookStatisticEntity>): List<Long>
 
   @Upsert
-  suspend fun update(bookStatistic: BookStatisticEntity): Long
+  suspend fun upsert(bookStatistic: BookStatisticEntity): Long
 
   @Query("SELECT * FROM book_statistic WHERE id = :id")
   suspend fun getById(id: BookId): BookStatisticEntity?
@@ -31,11 +31,13 @@ interface BookStatisticDao {
   @Query("SELECT * FROM book_statistic WHERE id IN (:bookIds)")
   suspend fun getByIds(bookIds: List<BookId>): List<BookStatisticEntity>
 
+  @Deprecated("")
   @Query("SELECT * FROM book_statistic WHERE id = :bookId")
   suspend fun getBookStatisticWithBookById(bookId: BookId): BookStatisticWithBookEntity?
 
+  @Deprecated("")
   @Query("SELECT * FROM book_statistic bs WHERE bs.priority > 0")
-  suspend fun getAllActive():List<BookStatisticWithBookEntity>
+  suspend fun getAllActive(): List<BookStatisticWithBookEntity>
 
   @Query("SELECT count(id) FROM book_statistic")
   suspend fun count(): Int
@@ -46,8 +48,19 @@ interface BookStatisticDao {
   @Query("DELETE FROM book_statistic")
   suspend fun deleteAll()
 
-  // flows
+  @Query("SELECT * FROM book_statistic WHERE id = :id")
+  fun observeById(id: BookId): Flow<BookStatisticEntity?>
 
+  @Query(
+    """
+    SELECT * FROM book_statistic 
+    WHERE (:minPriority IS NULL OR priority >= :minPriority) AND (:maxPriority IS NULL OR priority <= :maxPriority)
+    ORDER BY priority DESC"""
+  )
+  fun observeAll(minPriority: Int? = null, maxPriority: Int? = null): Flow<List<BookStatisticEntity>>
+
+
+  @Deprecated("")
   @Transaction
   @Query("SELECT * FROM book_statistic ORDER BY id")
   fun getAllBookStatisticWithBookFlow(): Flow<List<BookStatisticWithBookEntity>>