+++ domain/src/commonMain/kotlin/io/github/alelk/pws/domain/book/repository/BookRepository.kt
@@ -8,8 +8,8 @@ import io.github.alelk.pws.domain.core.ids.BookId
 import kotlinx.coroutines.flow.Flow
 
 interface BookRepository {
-  fun observeBook(id: BookId): Flow<BookDetail?>
-  fun observeBooks(query: BookQuery = BookQuery.Empty, sort: BookSort = BookSort.ByPriorityDesc): Flow<List<BookSummary>>
+  fun observe(id: BookId): Flow<BookDetail?>
+  fun observeMany(query: BookQuery = BookQuery.Empty, sort: BookSort = BookSort.ByPriorityDesc): Flow<List<BookSummary>>
 
   suspend fun get(id: BookId): BookDetail?
 }